Refactor plotting API to use Result objects
Updated all plotting functions to accept and validate Result objects instead of raw DataFrames, improving type safety and integration with the algorithm module. Added index/cols arguments for feature and interaction plots, improved error handling, and optimized internal data extraction. Also included performance optimizations in algorithm/utils.pyx for 3- and 4-column cases, and fixed float32 casting in plot_utils infinity replacement.

+++ treemind/plot/plot_utils.py
@@ -53,7 +53,7 @@ def _replace_infinity(
                 max_value * 0.1
             )  # Use 10% of max_value if only one unique point
         # Replace positive infinity values
-        data.loc[np.isposinf(data[column_name]), column_name] = max_value + difference
+        data.loc[np.isposinf(data[column_name]), column_name] = np.float32(max_value + difference)
 
     elif infinity_type == "negative":
         # Handle negative infinity
@@ -66,7 +66,7 @@ def _replace_infinity(
                 abs(min_value) * 0.1
             )  # Use 10% of min_value if only one unique point
         # Replace negative infinity values
-        data.loc[np.isneginf(data[column_name]), column_name] = min_value - difference
+        data.loc[np.isneginf(data[column_name]), column_name] = np.float32(min_value - difference)
 
     else:
         raise ValueError("infinity_type must be 'positive' or 'negative'")
